issue-614 move share intent receiving to ScanActivity
