+++ app/src/main/java/protect/card_locker/Utils.java
@@ -120,12 +120,7 @@ public class Utils {
 
             Bitmap bitmap;
             try {
-                Uri data;
-                if (requestCode == Utils.BARCODE_IMPORT_FROM_IMAGE_FILE) {
-                    data = intent.getData();
-                } else {
-                    data = intent.getParcelableExtra(Intent.EXTRA_STREAM);
-                }
+                Uri data = intent.getData();
                 if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) {
                     ImageDecoder.Source image_source = ImageDecoder.createSource(context.getContentResolver(), data);
                     bitmap = ImageDecoder.decodeBitmap(image_source, (decoder, info, source) -> decoder.setMutableRequired(true));
